SIHSort API

Sampling with interpolated histograms sorting algorithm, or SIHSort (pronounce sigh sort).

MPISort.mpisort! — Function

function mpisort!(
    v::AbstractVector;
    alg::SIHSort,
    lt=isless,
    by=identity,
    rev::Bool=false,
    order::Ordering=Forward,
)

Distributed MPI-based sorting API with the same inputs as the base Julia sorters.

Important: the input vector will be mutated, but the sorted elements for each MPI rank will be returned; this is required as the vector size will change with data migration.

MPISort.SIHSort — Type

struct SIHSort <: Base.Sort.Algorithm

Sampling with interpolated histograms sorting algorithm, or SIHSort (pronounce sigh sort).

Methods

SIHSort(comm)
SIHSort(comm, sorter)
SIHSort(;comm=MPI.COMM_WORLD, sorter=nothing, stats=SIHSortStats())

Fields

comm::MPI.Comm

: MPI communicator used. Default: MPI.COMM_WORLD

sorter::Union{Nothing, Function, Base.Sort.Algorithm}

: Local in-place sorter used. Default: nothing

stats::SIHSortStats

: Useful stats saved after sorting, e.g. elements' partitioning. Default: SIHSortStats()

MPISort.SIHSortStats — Type

mutable struct SIHSortStats

Useful stats saved after sorting.

Methods

SIHSortStats(splitters, num_elements)
SIHSortStats(;splitters=nothing, num_elements=nothing)

Fields

splitters::Union{Nothing, Vector}

: Values used to split elements across MPI ranks, length=nranks - 1 Default: nothing

num_elements::Union{Nothing, Vector{Int64}}

: Number of elements saved locally to each MPI rank, length=nranks. Default: nothing